[interfaz]

+++ EsqueletoJson/src/esqueletojson/Json.java
@@ -7,10 +7,16 @@ package esqueletojson;
 
 import Logica.Lista;
 import java.io.File;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.util.Iterator;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
+import org.json.simple.parser.ParseException;
 
 /**
  *
@@ -18,18 +24,18 @@ import org.json.simple.JSONObject;
  */
 public class Json {
 
-
     private Atributo atributos;
     private final String carpeta;
-    private Lista<Atributo> listaAtributos;
+    private final Lista<Atributo> listaAtributos;
+   // private  String nombrejson;
     
-
     public Json(String Carpet) {
         this.carpeta = Carpet;
         atributos = new Atributo(carpeta);
-        
-        
+        listaAtributos = new Lista();
     }
+    
+    
 
     /**
      * Agregar Json agrega json en la carpeta donde se encuentre la metadata y
@@ -38,44 +44,91 @@ public class Json {
      * @param json
      */
     public void AgregarJson(String json) {
+       
+        File directorio = new File("data/" + carpeta + "/" + json + ".json");
         JSONArray listatributos = new JSONArray();
         JSONObject obj = new JSONObject();
-        obj.put("Atributos", listatributos);
-        obj.put("nombre", json);
-        try {
-            FileWriter file = new FileWriter("data/" + carpeta + "/" + json + ".json");
-            file.write(obj.toString());
-            file.flush();
-            //Nodo<String> gato = listaCarpeta.Buscar(carpeta);
-        } catch (IOException e) {
-        }
+        if (directorio.exists()) {
 
+        } else {
+            obj.put("Atributos", listatributos);
+            obj.put("nombre", json);
+            try {
+                FileWriter file = new FileWriter("data/" + carpeta + "/" + json + ".json");
+                file.write(obj.toString());
+                file.flush();
+
+            } catch (IOException e) {
+            }
+        }
     }
+
     /**
-     * Agrega los atributos al json que faltan 
+     * Agrega los atributos al json que faltan
+     *
      * @param nameC
-     * @param document
+     * @param nombrejson
      * @param nombr
      * @param valo
      * @param tip
      * @param llav
      * @param requerid
      */
-    public void AgregarAtributos( String nameC,String document,String nombr, String valo, Tipo tip, Llave llav, boolean requerid) {
-        atributos = new Atributo(nameC,document, nombr,valo,tip,llav,requerid);
-        listaAtributos.Insertar(atributos);
+    public void AgregarAtributos(String nameC,String nombrejson, String nombr, String valo, Tipo tip, Llave llav, boolean requerid) {
+        if (listaAtributos.Existe(atributos)) {
+
+        } else {
+            System.out.println("entre aqui");
+            atributos = new Atributo(nameC,nombrejson, nombr, valo, tip, llav, requerid);
+            listaAtributos.Insertar(atributos);
+            
+            listaAtributos.Imprimir();
+        }
     }
+
+    public void CargarLista(String nombrejson) {
+        String path = "data/" + carpeta + "/"+ nombrejson+".json";
+        JSONParser parser = new JSONParser();
+        FileReader fr = null;
+        try {
+            fr = new FileReader(path);
+        } catch (Exception e) {
+            try {
+                File f = new File(path);
+                f.createNewFile();
+            } catch (IOException ex) {
+                Logger.getLogger(Documentos.class.getName()).log(Level.SEVERE, null, ex);
+            }
+        }
+        try {
+            Object obj = parser.parse(fr);
+            JSONObject jsonObjeto = (JSONObject) obj;
+            JSONArray courseArray = (JSONArray) jsonObjeto.get("Atributos");
+            Iterator<Atributo> iterator = courseArray.iterator();
+            while (iterator.hasNext()) {
+
+                listaAtributos.Insertar(iterator.next());
+            }
+            //listaJson.Imprimir();
+
+        } catch (IOException | ParseException ex) {
+            Logger.getLogger(Documentos.class.getName()).log(Level.SEVERE, null, ex);
+        }
+
+    }
+
     /**
      * Guardas los atributos que estan las listas y los escribe en el json
      */
     public void Commit() {
-      atributos.InsertarAtributo();
+        atributos.InsertarAtributo();
     }
-    
+
     /**
-    *Busca los json existentes en la carpeta
-    * @param json
-    */
+     * Busca los json existentes en la carpeta
+     *
+     * @param json
+     */
     public void BuscarJson(String json) {
         File directorio = new File("data/" + carpeta + "/" + json + ".json");
         if (directorio.exists()) {
@@ -84,12 +137,17 @@ public class Json {
             System.out.println("no existe");
         }
     }
+
     /**
      * Busca el elemento dado y lo elimina
-     * @param json 
+     *
+     * @param json
      */
     public void EliminarJson(String json) {
+        String path = "data/" + carpeta + "/metadata.json";
         File directorio = new File("data/" + carpeta + "/" + json + ".json");
+        JSONParser parser = new JSONParser();
+        FileReader fr = null;
         if (directorio.exists()) {
             try {
                 directorio.deleteOnExit();
@@ -98,6 +156,33 @@ public class Json {
             }
         } else {
             System.out.println("no existe");
+        }
+        try {
+            fr = new FileReader(path);
+        } catch (Exception e) {
+            try {
+                File f = new File(path);
+                f.createNewFile();
+            } catch (IOException ex) {
+                Logger.getLogger(Metadata.class.getName()).log(Level.SEVERE, null, ex);
+            }
+
+        }
+        try {
+            Object obj = parser.parse(fr);
+            JSONObject jsonObjeto = (JSONObject) obj;
+
+            JSONArray jsonArray = (JSONArray) jsonObjeto.get("DocumentosJson");
+            jsonArray.remove(json);
+            jsonObjeto.put("DocumentosJson", jsonArray);
+            try (FileWriter file = new FileWriter(path)) {
+                file.write(obj.toString());
+                file.flush();
+            } catch (IOException e) {
+            }
+        } catch (ParseException ex) {
+        } catch (IOException ex) {
+            Logger.getLogger(Metadata.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
